# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Banco de Horas Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Criar um novo usuário administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Usuário já existe ou dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Autenticar usuário com usuário e senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Autenticado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Usuário ou senha incorretos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Nome de usuário para acesso
        password:
          type: string
          description: Senha de acesso
          format: password
    AuthResponse:
      type: object
      required: [success, message, username]
      properties:
        success:
          type: boolean
        message:
          type: string
        username:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string